backend/pyproject.toml
[project]
name = "datax-backend"
version = "0.1.0"
description = "DataX - AI-native data analytics platform backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "pydantic-ai>=0.1,<1",
    "sqlalchemy>=2.0,<3",
    "alembic>=1.14,<2",
    "duckdb>=1.2,<2",
    "cryptography>=44,<45",
    "uvicorn[standard]>=0.34,<1",
    "sse-starlette>=2.0,<3",
    "structlog>=25,<26",
    "psycopg[binary]>=3.2,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.28,<1",
    "ruff>=0.9,<1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
